# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Slooza Assignment Backend API
  description: API documentation for the Slooza Assignment Backend
  version: 0.0.1
servers:
  - url: http://localhost:3000
    description: Local server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginDto:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    CreateOrderDto:
      type: object
      properties:
        restaurantId:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemDto'
      required:
        - restaurantId
        - items

    OrderItemDto:
      type: object
      properties:
        menuItemId:
          type: string
        quantity:
          type: integer
          minimum: 1
      required:
        - menuItemId
        - quantity

    PlaceOrderDto:
      type: object
      properties:
        address:
          type: string
        paymentMethod:
          type: string
          enum: [CASH, CARD, UPI]
      required:
        - address

    Restaurant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: string
        country:
          type: string
        menuItems:
          type: array
          items:
            $ref: '#/components/schemas/MenuItem'

    MenuItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        restaurantId:
          type: string

    Order:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        restaurantId:
          type: string
        status:
          type: string
          enum: [PENDING, CONFIRMED, PREPARING, READY, COMPLETED, CANCELLED]
        totalAmount:
          type: number
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        payment:
          $ref: '#/components/schemas/Payment'

    OrderItem:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        menuItemId:
          type: string
        quantity:
          type: integer
        price:
          type: number

    Payment:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        method:
          type: string
          enum: [CASH, CARD, UPI]
        status:
          type: string
          enum: [PENDING, SUCCESS, FAILED]
        amount:
          type: number

paths:
  /:
    get:
      summary: Health check
      tags:
        - App
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
                example: Hello World!

  /auth/login:
    post:
      summary: User login
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
        '401':
          description: Unauthorized

  /restaurants:
    get:
      summary: Get all restaurants
      tags:
        - Restaurants
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Items per page
        - in: query
          name: search
          schema:
            type: string
          description: Search by name
        - in: query
          name: dietary
          schema:
            type: string
          description: Filter by dietary preference
      responses:
        '200':
          description: List of restaurants
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: 
                    type: array
                    items:
                      $ref: '#/components/schemas/Restaurant'
                  meta:
                    type: object

  /restaurants/{id}:
    get:
      summary: Get a restaurant by ID
      tags:
        - Restaurants
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Restaurant ID
      responses:
        '200':
          description: Restaurant details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        '404':
          description: Restaurant not found

  /restaurants/{id}/menu:
    get:
      summary: Get restaurant menu
      tags:
        - Restaurants
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Restaurant ID
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Items per page
        - in: query
          name: search
          schema:
            type: string
          description: Search by name
        - in: query
          name: dietary
          schema:
            type: string
          description: Filter by dietary preference
      responses:
        '200':
          description: Restaurant menu items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/MenuItem'
                  meta:
                    type: object

  /orders:
    post:
      summary: Create a new order
      tags:
        - Orders
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderDto'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
    get:
      summary: Get all orders
      tags:
        - Orders
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: country
          schema:
            type: string
          required: false
          description: Filter by country (Admin only)
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'

  /orders/{id}:
    get:
      summary: Get an order by ID
      tags:
        - Orders
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Order ID
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found

  /orders/{id}/place:
    put:
      summary: Place an order
      tags:
        - Orders
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Order ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceOrderDto'
      responses:
        '200':
          description: Order placed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

  /orders/{id}/cancel:
    put:
      summary: Cancel an order
      tags:
        - Orders
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Order ID
      responses:
        '200':
          description: Order cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

  /payments/checkout:
    post:
      summary: Process payment for an order
      tags:
        - Payments
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderId:
                  type: string
                method:
                  type: string
                  enum: [CASH, CARD, UPI]
              required:
                - orderId
                - method
      responses:
        '200':
          description: Payment processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
                    enum: [PENDING, SUCCESS, FAILED]

  /payments/{orderId}:
    put:
      summary: Update payment method (Admin only)
      tags:
        - Payments
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: orderId
          schema:
            type: string
          required: true
          description: Order ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                method:
                  type: string
                  enum: [CASH, CARD, UPI]
              required:
                - method
      responses:
        '200':
          description: Payment method updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'